requirements.txt
pandas
python-dotenv
plotly
kaleido
numpy
matplotlib
requests


